Explorations of unboxed sum type Or.  Very successful explorations!
Mapping can get near-handrolled speed, save for primitives which are boxed.
The biggest win of all would be if primitives could be unboxed, but that
is impossible without compiler support.  Even so, you can save a fair bit
of memory churn and a precious nanosecond or two by avoiding boxing.