Fixed strange behavior in regularizedIncompleteBeta.
When a large, b=1, x close to 1, we'd be off by a factor of 10.

This didn't happen when b was a tiny bit above or below 1.

Switching abm < 1 to abm <= 1 seems to have solved the problem, though I was not able to identify where it was going wrong to begin with.